Adicionado coisas taredas
