+++ React/src/components/Header.js
@@ -1,12 +1,18 @@
-import React from 'react';
+import React, {useState} from 'react';
 import Navbar from './Navbar'
 import Sidebar from './Sidebar'
 
 const Header = () => {
+    const [name, setName] = useState('');
+
+    const getName = (newName) => {
+        setName(newName);
+    }
+
     return (
         <div className="Header">
-            <Navbar />
-            <Sidebar />
+            <Navbar name={name} />
+            <Sidebar getName={getName} />
         </div>
     )
 }
