+++ React/src/components/Navbar.js
@@ -3,13 +3,13 @@ import AppBar from '@material-ui/core/AppBar'
 import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
 
-const NavBar = () => {
+const NavBar = (props) => {
     return (
         <div>
             <AppBar position="fixed" className="Navbar">
                 <Toolbar>
                     <Typography variant="title" color="inherit">
-                        About My Name
+                        About My Name: {props.name}
                     </Typography>
                 </Toolbar>
             </AppBar>
