+++ React/src/components/Sidebar.js
@@ -1,11 +1,12 @@
 import React from 'react'
 import Container from '@material-ui/core/Container'
 import FindName from './FindName'
+import { Typography } from '@material-ui/core';
 
-const Sidebar = () => {
+const Sidebar = (props) => {
     return (
         <Container className="ContainerSidebar">
-            <FindName />
+            <FindName getName={props.getName} />
         </Container>
     )
 }
